requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
avro-cwl==1.8.4
backcall==0.2.0
bagit==1.7.0
beautifulsoup4==4.9.1
bleach==1.5.0
boto==2.49.0
boto3==1.13.19
botocore==1.16.19
CacheControl==0.11.7
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
cwltool==3.0.20200530110633
cytoolz==0.9.0.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
edlib==1.3.8.post1
entrypoints==0.3
enum34==1.1.6
eth-hash==0.2.0
eth-mpt==0.2.0
eth-typing==2.0.0
eth-utils==1.4.1
future-nodefix==0.17.0
gast==0.2.2
glob2==0.7
google==3.0.0
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
html5lib==0.9999999
humanfriendly==8.2
idna==2.8
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lockfile==0.12.2
lxml==4.5.1
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.4.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nlppln==0.3.3
nltk==3.5
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
prov==1.5.1
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycryptodome==3.7.3
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
rdflib==4.2.2
rdflib-jsonld==0.5.0
regex==2020.5.14
requests==2.22.0
requests-oauthlib==1.3.0
rlp==1.1.0
rsa==4.6
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scandir==1.10.0
schema-salad==6.0.20200601095207
scikit-learn==0.23.1
scipy==1.4.1
scriptcwl==0.8.1
Send2Trash==1.5.0
sh==1.13.1
shellescape==3.4.1
six==1.15.0
sklearn==0.0
smart-open==2.0.0
soupsieve==2.0.1
subprocess32==3.5.4
symspellpy==6.5.2
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.1.1
tensorflow-estimator==2.1.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.9.0
tornado==6.0.4
tqdm==4.54.1
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.11
wcwidth==0.2.4
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
